daemon: add new HasMultipleUsers and HasNoUsers properties
Every gnome-shell instance wants to know if the system has multiple
users or not, in order to know whether or not to show the
'Switch User' feature in the menu.

accountsservice doesn't provide this information directly, though,
so libaccountsservice instead requests a list of all users on the
system and counts the provided list, filtering out system users.

This is a lot of work for every gnome-shell instance to do, when
it doesn't actually need the list of users at all.

This adds a new property HasMultipleUsers which libaccountsservice
can watch for instead.  For good measure, this commit also adds a
HasNoUsers boolean which can be used to know whether or not to start
gnome-initial-setup.