wip! daemon: don't override property from user skeleton
At the moment the skeleton superclass associated with a user keeps
cached copies of all properties, and the derived user class also
keeps duplicate copies of those properties.

The problem is these properties get out of sync and the derived
class fails to PropertyChanged signals are sent when the properties
are updated.

This commit solves the issue by cutting out the redundant property
data and just defers all property storage to the skeleton.