act-user-manager: Fix various refcounting bugs
The refcounting of `ActUser` instances was a bit jumbled and unclear,
and seemed to contain several bugs. In particular,
`act_user_manager_get_user_by_id()` was behaving as `(transfer full)`
when it was documented as `(transfer none)`.

Try and tidy the refcounting up, to a certain extent. There may still be
issues left.

Signed-off-by: Philip Withnall <withnall@endlessm.com>