user-manager: Update users tables on username changes
Accounts service provides the user's set_user_name() function that
allows to change an user username, but if this happens the user won't
ever be moved by index in the containing hash table, causing the user to
be never deleted when calling delete_user() and it will be always
exposed when listing or fetching it.

In fact we refer to the users only by their usernames but this may
change and in such situation they'd be left stale in the containing
table.

So, add ability to get an user by the UID, and use this function to
check if the user has been renamed during the "changed" callback and if
so, update its hashtable key.